# Speed up complex dependency resolves by creating DepProxy factory and cache