--- a/bundler/spec/bundler/dep_proxy_spec.rb
+++ b/bundler/spec/bundler/dep_proxy_spec.rb
@@ -2,10 +2,10 @@
 
 RSpec.describe Bundler::DepProxy do
   let(:dep) { Bundler::Dependency.new("rake", ">= 0") }
-  subject { described_class.new(dep, Gem::Platform::RUBY) }
+  subject { described_class.get_proxy(dep, Gem::Platform::RUBY) }
   let(:same) { subject }
-  let(:other) { subject.dup }
-  let(:different) { described_class.new(dep, Gem::Platform::JAVA) }
+  let(:other) { described_class.get_proxy(dep, Gem::Platform::RUBY) }
+  let(:different) { described_class.get_proxy(dep, Gem::Platform::JAVA) }
 
   describe "#eql?" do
     it { expect(subject.eql?(same)).to be true }
@@ -15,8 +15,9 @@ RSpec.describe Bundler::DepProxy do
     it { expect(subject.eql?("foobar")).to be false }
   end
 
-  describe "#hash" do
-    it { expect(subject.hash).to eq(same.hash) }
-    it { expect(subject.hash).to eq(other.hash) }
+  describe "must use factory methods" do
+    it { expect { described_class.new(dep, Gem::Platform::RUBY) }.to raise_error NoMethodError }
+    it { expect { subject.dup }.to raise_error NoMethodError }
+    it { expect { subject.clone }.to raise_error NoMethodError }
   end
 end