--- a/bundler/spec/resolver/basic_spec.rb
+++ b/bundler/spec/resolver/basic_spec.rb
@@ -176,7 +176,7 @@ Bundler could not find compatible versions for gem "a":
 
     deps = []
     @deps.each do |d|
-      deps << Bundler::DepProxy.new(d, "ruby")
+      deps << Bundler::DepProxy.get_proxy(d, "ruby")
     end
 
     should_resolve_and_include %w[foo-1.0.0 bar-1.0.0], [[]]