--- a/bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/bson/GET/request
+++ b/bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/bson/GET/request
@@ -1,0 +1,7 @@
+> GET /info/bson
+> accept-encoding: gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+> accept: */*
+> user-agent: bundler/2.3.0.dev rubygems/3.3.0.dev ruby/2.7.2 (x86_64-pc-linux-gnu) command/install options/no_install,retry,spec_run 4c3f19426c5c0122
+> connection: keep-alive
+> keep-alive: 30
+> host: index.rubygems.org