--- a/bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/durran-validatable/GET/response
+++ b/bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/durran-validatable/GET/response
@@ -1,0 +1,37 @@
+HTTP/1.1 200 OK
+connection: keep-alive
+content-length: 652
+content-type: text/plain; charset=utf-8
+x-frame-options: SAMEORIGIN
+x-xss-protection: 1; mode=block
+x-content-type-options: nosniff
+x-download-options: noopen
+x-permitted-cross-domain-policies: none
+referrer-policy: strict-origin-when-cross-origin
+etag: "952c7172cd0bce742f794d16c2b78260"
+last-modified: Mon, 14 Dec 2020 15:03:53 GMT
+cache-control: max-age=60, public
+x-request-id: 4432ad5f-8194-445d-a1eb-48cb2c28ec22
+x-runtime: 0.006949
+strict-transport-security: max-age=31536000
+x-backend: F_Rails 54.214.179.27:443
+accept-ranges: bytes
+date: Thu, 07 Jan 2021 17:55:42 GMT
+via: 1.1 varnish
+age: 3416
+x-served-by: cache-mad22047-MAD
+x-cache: HIT
+x-cache-hits: 1
+x-timer: S1610042143.751850,VS0,VE0
+vary: Accept-Encoding,Fastly-SSL
+server: RubyGems.org
+
+---
+1.7.5 |checksum:937368bad4d2edce9bafb215cafb6baa42f61fd5ee421b59eb4567c44956855b
+1.8.0 |checksum:b20330beb2900c2a2bbb9e6d0fb05e63d4ec72b5d22a60b12658cf9381b52ab6
+1.8.1 |checksum:6c3c232a5b616fe5d4a18cdd939632339a0f83183b03c1333cd86870b1b2e559
+1.8.2 |checksum:606ee919c3ce5ec6b6216d7f68e6a32ac2b7ace7361f4ed60e30ed247bc5d420
+1.8.3 |checksum:d36083eb616cced8963eae096062729a8d9b96216f69c2a98470afc93c8abd7e
+1.8.4 |checksum:2cfb8b793b85afe900fe57b0790ff9cdaf0fbec65be13f39e10713feaec42d02
+2.0.0 |checksum:c1cd755a2fa7bc7204f3a2f0f3b1d2327ae8bb1e1a3a123b61fa24020cfb707b
+2.0.1 |checksum:7dbcb3e418a0fd25e99fc20239e700979ceb60b4b3149d3d53044de48fa97e5d