--- a/bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/leshill-will_paginate/GET/response
+++ b/bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/leshill-will_paginate/GET/response
@@ -1,0 +1,30 @@
+HTTP/1.1 200 OK
+connection: keep-alive
+content-length: 86
+content-type: text/plain; charset=utf-8
+x-frame-options: SAMEORIGIN
+x-xss-protection: 1; mode=block
+x-content-type-options: nosniff
+x-download-options: noopen
+x-permitted-cross-domain-policies: none
+referrer-policy: strict-origin-when-cross-origin
+etag: "66eeb8db4f23eaeb236a1beb9f69a4da"
+last-modified: Mon, 14 Dec 2020 15:02:35 GMT
+cache-control: max-age=60, public
+x-request-id: 9a032d13-ff2c-4503-8d08-4d1828297263
+x-runtime: 0.007636
+strict-transport-security: max-age=31536000
+x-backend: F_Rails 44.240.15.233:443
+accept-ranges: bytes
+date: Thu, 07 Jan 2021 17:55:42 GMT
+via: 1.1 varnish
+age: 3416
+x-served-by: cache-mad22020-MAD
+x-cache: HIT
+x-cache-hits: 1
+x-timer: S1610042143.772021,VS0,VE0
+vary: Accept-Encoding,Fastly-SSL
+server: RubyGems.org
+
+---
+2.3.11 |checksum:cfdc490d2d97ee782cd8a546c9d9d246b4085e735635796673d94ee4859ccd91