--- a/bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/mislav-will_paginate/GET/response
+++ b/bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/mislav-will_paginate/GET/response
@@ -1,0 +1,40 @@
+HTTP/1.1 200 OK
+connection: keep-alive
+content-length: 962
+content-type: text/plain; charset=utf-8
+x-frame-options: SAMEORIGIN
+x-xss-protection: 1; mode=block
+x-content-type-options: nosniff
+x-download-options: noopen
+x-permitted-cross-domain-policies: none
+referrer-policy: strict-origin-when-cross-origin
+etag: "b1628db48f70a3ec7444a66171b9045f"
+last-modified: Mon, 14 Dec 2020 15:54:46 GMT
+cache-control: max-age=60, public
+x-request-id: ae277e74-ed96-4c76-b808-bb27d7a39577
+x-runtime: 0.007963
+strict-transport-security: max-age=31536000
+x-backend: F_Rails 54.214.179.27:443
+accept-ranges: bytes
+date: Thu, 07 Jan 2021 17:55:42 GMT
+via: 1.1 varnish
+age: 3416
+x-served-by: cache-mad22069-MAD
+x-cache: HIT
+x-cache-hits: 1
+x-timer: S1610042143.777695,VS0,VE0
+vary: Accept-Encoding,Fastly-SSL
+server: RubyGems.org
+
+---
+2.3.8 |checksum:1b78aae7894093ab00122476f437403502ef3301b428e37cbabfd44dfd083d81
+2.3.7 |checksum:ef70a5a3b277114fa9fd2ebe7a6bda15b546483182437cb1db6a1f7eb2c464e3
+2.3.6 |checksum:c62f08034cad595124f192891423db5dbf49597d81aa6dbc9c9936ebda2de947
+2.3.5 activesupport:>= 1.4.4|checksum:a1c6b18b61a668bd3b0fee183a82c229e04f23f59f5c55d8e9ee5f50f75bc122
+2.3.4 |checksum:d8ecbec250b0c83982d17281fdbbbca228c46d1445553bd747a7b80de6a095f5
+2.3.3 activesupport:>= 1.4.4|checksum:d18e5693ebe90e2049be82d49a94e5e230e7f7144ef33ffb3684fe36f5160906
+2.3.2 activesupport:>= 1.4.4|checksum:2b48a820024336bb2d7a9971cba153a8e83e46680d11c1bb57e5720c53239536
+2.3.10 |checksum:bb366303625d54c51fc2c4a31e031584b42e7d08d69a4ebb8a3a8cce921f59b4
+2.3.1 |checksum:bf84377fd658d907e7ad4532f2d7369a193b04079e5cbb1bdaf3fb158ea9ba1c
+2.3.0 |checksum:1178c63cea3b70377e9f1d940c628c1a310e2240f42ac23722e9805db5b39983
+2.2.3 |checksum:e625b0e3ab8e6ec380e29688644d5fcfb211268a225be8f20bd08b0f6ecff52a