--- a/bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/mongo_ext/GET/response
+++ b/bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/mongo_ext/GET/response
@@ -1,0 +1,42 @@
+HTTP/1.1 200 OK
+connection: keep-alive
+content-length: 1072
+content-type: text/plain; charset=utf-8
+x-frame-options: SAMEORIGIN
+x-xss-protection: 1; mode=block
+x-content-type-options: nosniff
+x-download-options: noopen
+x-permitted-cross-domain-policies: none
+referrer-policy: strict-origin-when-cross-origin
+etag: "c91397b5338a81744314015e58090e13"
+last-modified: Mon, 14 Dec 2020 15:05:18 GMT
+cache-control: max-age=60, public
+x-request-id: b586dc3e-8b19-4837-97c3-7033d5ea2598
+x-runtime: 0.006655
+strict-transport-security: max-age=31536000
+x-backend: F_Rails 44.240.15.233:443
+accept-ranges: bytes
+date: Thu, 07 Jan 2021 17:55:42 GMT
+via: 1.1 varnish
+age: 3416
+x-served-by: cache-mad22078-MAD
+x-cache: HIT
+x-cache-hits: 1
+x-timer: S1610042143.782212,VS0,VE0
+vary: Accept-Encoding,Fastly-SSL
+server: RubyGems.org
+
+---
+0.15 mongo:= 0.15|checksum:bfd847ea8c2983efe9ff150669dbac37a17a7ac985bbb6d32a0327af2b6ddca9
+0.15.1 |checksum:0e7a2259930c9c2261dcc0819e17bf70ab89bacbc33d587b5f9c6410f2b8d268
+0.16 |checksum:49e2224cec622832c8e0035bf61b7cadc5d78fc6c9dac4bfd7b00b53bdfa4324
+0.17 |checksum:27eb2329dba85aacb0ba0fb22952d08ebdb1c518d4344e367fec9d16732ba5bb
+0.17.1 |checksum:d209b50264081cd227af64dda664338b536111949eeccd4b780bca6237c034d1
+0.18 |checksum:09b95a41c189b7278045b2d8f9b94498652da5cd432350ca7846ec4214c9bb9e
+0.18.1 |checksum:08fe4cacedcf4bbbd51a588972406ae2dd386e84e093507b081497c46f9284b0
+0.18.2 |checksum:30107c7d0648b52d9d14ed4fd91383a5335c2a5b920ee34e805662b8d9261429
+0.18.3 |checksum:0862536ef8ef777142e84851a1465d5e658c5a65007c19a1a6fd042a2f25ae17
+0.19 |checksum:f25af22d6f87ec73c45f9a3e6b3af732dc759fb161fa68d871090b509f644a16
+0.19.1 |checksum:89482d7eeb8b86b601cee3b5ac7f76fcaa7fd99cbe580122df6698ee6c150cd0
+0.19.2 |checksum:2f683475ef5b90f3050a4ae17c000915fe3df9af6e66cb976fdfe24ea860d6b1
+0.19.3 |checksum:d686eb3010cf83b35a2a49a8e4765709019f385da10fb4a328db054c6573698a