--- a/bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/mongodb-mongo/GET/response
+++ b/bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/mongodb-mongo/GET/response
@@ -1,0 +1,70 @@
+HTTP/1.1 200 OK
+connection: keep-alive
+content-length: 3317
+content-type: text/plain; charset=utf-8
+x-frame-options: SAMEORIGIN
+x-xss-protection: 1; mode=block
+x-content-type-options: nosniff
+x-download-options: noopen
+x-permitted-cross-domain-policies: none
+referrer-policy: strict-origin-when-cross-origin
+etag: "b92d59af7bf2e217ecfa4f6f5e945682"
+last-modified: Mon, 14 Dec 2020 15:54:44 GMT
+cache-control: max-age=60, public
+x-request-id: 4a88d671-08ae-45d8-943a-2cde527c8b45
+x-runtime: 0.007970
+strict-transport-security: max-age=31536000
+x-backend: F_Rails 44.240.15.233:443
+accept-ranges: bytes
+date: Thu, 07 Jan 2021 17:55:42 GMT
+via: 1.1 varnish
+age: 3416
+x-served-by: cache-mad22035-MAD
+x-cache: HIT
+x-cache-hits: 1
+x-timer: S1610042143.772162,VS0,VE1
+vary: Accept-Encoding,Fastly-SSL
+server: RubyGems.org
+
+---
+0.9 |checksum:38e2f55e3f51bc5c01aea7d43ada2f743f38c36410df945e42be73e98f655e91
+0.8 |checksum:5f63e9f4c01bf165254886f90321e9c3e0062416646f50bba52096e26b88e22c
+0.7 |checksum:81dc0faf59ca517bb09e6931def5ac2f7df1cca2231717bd8aab5bbe4e940308
+0.6.7 |checksum:7ebcdf9cfc0184221a1058318d9b960c588a2d7f1f5c5b01b5b8900f8f42005a
+0.6.6 |checksum:abf23b5037e7c927fe0565ab26048d01e2d1688456de0e573dddbff786cb4e46
+0.6.5 |checksum:3c3b20f926d59125bbc01ce121ace88cf74a38287fad0ee9590d7f97147116a6
+0.6.4 |checksum:fa21d4fb4c116d6031dd778ca59a55bb14de805def0b114eea7521c15ff71d66
+0.6.3 |checksum:ce3919ae3500ee1740904ac8360de1f6b128f87244304bbb5835baeb27570604
+0.6.2 |checksum:0f3b80aafb1c778affc8a922a682b7e43af822668266b2198f43d9a9bdc5bc54
+0.6.1 |checksum:d62d84fd9b88c93c6b0212d14ecb89b33413d8d0650471d0ce3b4c84ee5fd2ab
+0.5.3 |checksum:770e97ff7db143fda67dda5d2f121dc5b4edff6d186bc3431e99c9a09372d813
+0.5.2 |checksum:5b66149a9c3797e75025d9e02bb31f3f330f396c18ddf9b660162ff1266dccaf
+0.5.1 |checksum:ce1bc9ebe8557040255ae7c23d0261e10083be701b9dfa73b7701dbd41e7d36e
+0.5.0 |checksum:7705ea2aad5f6836586cdb3b0ae3a5ea112c55f9365144d83d720f0e048f4096
+0.4.4 |checksum:42d537ce05df3317a71464fa2b56adaf68f89708fda3f67bce5f748f4a0d49e7
+0.4.3 |checksum:5ddc35fb614e0fc464698624433b0425c0b1c69f4ad4b7ea12dd261e4e668fd7
+0.4.2 |checksum:0a2cbf2e3069c17d17cb4c7e78ac333db0503a877d83334497e2aaaf3a39aa18
+0.4.1 |checksum:a8c5cdd3a61aa6ab5966612f9da2c4f661c8d933b149ba04bdd0a3c98575f3ca
+0.4.0 |checksum:cfa2f0db1f2fed66606354acf3327e0efe1dc30845936d0668a10dcecafb9251
+0.3.3 |checksum:da6ee0058d84abd00b1b3e5b73e56839a13a351ff60852a54664551a403b0869
+0.3.2 |checksum:600a0aa5b899527bedbbdbba31078cbdd8ef24e67a5fee7c6ac72f88c001191d
+0.3.1 |checksum:8a69a3e338d10113f5f200b06cdc03c9c16cb55ab2d3498158092e1f4aa13266
+0.3.0 |checksum:1f11e4a955919d32af6f025750cd6a1c275032143b11b57fea4d5242c24bcd59
+0.2.7 |checksum:88f2b3f72343509f47f51d90d361abc9c9b59e1f352870e3906619c33183ef3b
+0.2.6 |checksum:6d23f774290fcca1e7af060fd6a89060e32871cccb4927c00050e957d9a1209d
+0.2.5 |checksum:f82fc3cfe9d75426cc68891996bf5624f280cbb39538d6143e12ea08384ad5da
+0.2.4 |checksum:553492323e2108dbfa4193d2a89df5c237f2540dc2f231fb1695c589e03245c5
+0.2.3 |checksum:b37cc890a9b5be924672d656c1fcab8e45801e44c4c7f8c51c57e797b8baa39b
+0.2.2 |checksum:4849df05ddffbcea7c38bc077ec2df82b2e7d97af8a0187ea85e1239f6fa5f22
+0.2.1 |checksum:16ce3385dbed4f14f1d93213e24e6d4a4c09ee96505b867718defead1fe693e9
+0.2.0 |checksum:e7980a5d5470de365319aa1437fba6f018e1aeb60869b3dc6f37e7cf623982da
+0.14 |checksum:c9ac3e2af88d954181b9842dc6b240599599f16dc3b2dc791bfb55f6b07974b8
+0.14.1 |checksum:39d056f693642150f8a56ba440488e38852d1e6252c70fa51a2ebf89c62e12d5
+0.13 |checksum:959c0bbae0ba0a459d0e6657be216ccc4447836a30bce0685cc4edc3ad5119b8
+0.12 |checksum:db4d31aad18625e6c018499837cc1e432ac741952840fb0d83254b24f16bb02f
+0.11 |checksum:77f4d679e7355ea06344c05854e418b9a048dd943102ac7bf1fff0251d1e501f
+0.11.1 |checksum:335f86c3a64772a44c3c3dc1f70cfbac3c522cc493ec32bae9838896b4f88bb5
+0.10 |checksum:c2d0f9d1a8fbde6323b5b5aaa36f5dbb791aa1abf4d7f9c191cfae3e9de941b1
+0.10.1 |checksum:0d4ce66e34cb08b1dad1959ab5cb6eace3ca7c89316a602dd28886c28cfb2222
+0.1.4 |checksum:3c337b1640e1f6284c98b2569178e527c744bcdfa1f0225c314be35d22260f21
+0.1.3 |checksum:008fc4fb242e0a74377696920c4a83d28f3ffb05fca958466575c63c6fe5673a