--- a/bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/moped/GET/response
+++ b/bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/moped/GET/response
@@ -1,0 +1,80 @@
+HTTP/1.1 200 OK
+connection: keep-alive
+content-length: 5217
+content-type: text/plain; charset=utf-8
+x-frame-options: SAMEORIGIN
+x-xss-protection: 1; mode=block
+x-content-type-options: nosniff
+x-download-options: noopen
+x-permitted-cross-domain-policies: none
+referrer-policy: strict-origin-when-cross-origin
+etag: "bb48b758f63a641e94c2bacab4fe12a6"
+last-modified: Mon, 14 Dec 2020 16:14:36 GMT
+cache-control: max-age=60, public
+x-request-id: fe63adbc-fe72-4c72-b41b-5d2185ff955a
+x-runtime: 0.009433
+strict-transport-security: max-age=31536000
+x-backend: F_Rails 44.237.26.162:443
+accept-ranges: bytes
+date: Thu, 07 Jan 2021 17:55:42 GMT
+via: 1.1 varnish
+age: 3417
+x-served-by: cache-mad22059-MAD
+x-cache: HIT
+x-cache-hits: 1
+x-timer: S1610042143.813226,VS0,VE1
+vary: Accept-Encoding,Fastly-SSL
+server: RubyGems.org
+
+---
+0.0.0.beta |checksum:6d008b19284ebb445ef1d07b1991a8e9ce8ae73ea46af42a6e32bce96a33e7f8,rubygems:>= 1.3.6
+1.0.0.alpha |checksum:90bda5c8050ffebaf5cf2e1c182d07bd842cd41e17e6205d3280023d833a05ca,rubygems:> 1.3.1
+1.0.0.beta |checksum:47c975d8cc27b217c22e69239a26f3e3a23bf5d338f4c4913f333bd24ffb1e7d,rubygems:> 1.3.1
+1.0.0.rc |checksum:8c763272fa0adc0aa8bd3a9ca4688c0462f1c6490a989fe537c08bd2f95aac3b,rubygems:> 1.3.1
+1.0.0 |checksum:2b05dc30c5e3953d5f41b9bc15df304027cf443f041249cb8c4608dd2ec05ebe
+1.1.0 |checksum:1331da5da2be919e7a9cf9cfbaae441088a6e49489949b91b216e02c5d4278df
+1.1.1 |checksum:a9e77a4ac7d46f266114c7b585a522f707d6e359f32429733fc9aab1fc05a9cb
+1.0.1 |checksum:576a2782139cdda40dd3af0420914ae0b5fa1c2302e126a882993ea913e2a4ab
+1.1.2 |checksum:7b3d9c7c026cb2b9d5f93e09680a5f1a24ccbc24012f84aa7a901b3fabac1220
+1.1.3 |checksum:95ba2d8172580977e51420d092547a7352f70ff2abfe8a225a76c8bf5c59d415
+1.1.4 |checksum:fd20e8334b1c8b350063743d4e3934f50311f4555dd316f2499671e885da53bd
+1.1.5 |checksum:d000f90219fd5bb537ccf3da127f11a538f9602b65238bf7b74475862714f725
+1.1.6 |checksum:80d92fcdf0926883b0e30b7fb9693a174a88c3f8be3fd4ab744e88fa8532f704
+1.2.0 |checksum:21d5a5c93e783b4de5d5317251fd4f47f0bdd8cf771f13200d12f92130957d45
+1.2.1 |checksum:2e3b4ae0bac0be67c81a9952aea7e9cee2339abc1dc883da7897510107e4eb83
+1.2.2 |checksum:e7e5116a20497acef2fe2240e395b6269cb8805a0fec0595fc33b163696f069a
+1.2.4 |checksum:b023f806ee6a8753ad963547a8d9426593bc049312857d6844e27eab467d4259
+1.2.5 |checksum:923904d8249d77252ed62cd3e94d91c66d145a8508332e1211dd86ab91750a51
+1.2.6 |checksum:a0e7ac70587784c485b140518601dfd2bbd3144b48f0072de1094e31787224ca
+1.2.7 |checksum:1a2a1ecb248bf0683bd1b7f062fdfedbcdd15d3e2e6884e96f345547a9d7041c
+1.2.8 |checksum:dc127037b3fe81a691d63f838f844feeb453f6cf53d913113acaa0f27250d47c
+1.2.9 |checksum:0b9a07c1c7eb4659c1f40e912c91def23715fba69b018c1a68341538604ba034
+1.3.0 |checksum:02d6a7da7b8d5618d697cc92cb1a7c0071e981de4d24eebbc5748f3ea646235d
+1.3.1 |checksum:b894e4d37a8e016e84fd7eec63f0b0e66897412ff72d331c52385403539da8ed
+1.3.2 |checksum:5300330422b97eddadd1629cd4d782e3c44d3308d0dcf9e587c11c71875b532c
+1.4.0 |checksum:2fe1fc8fe548934e57da9c0ecc33b049ba376ae84e3c9515ba477b5e90124818
+1.4.1 |checksum:14b692d9bd2a377ccdaf5c13bf9111738d619b8c59813b452827f180557bcb84
+1.4.2 |checksum:a25f1200e32bf4a7e674cb7b84e507af586aeac4e182870ae8981af42dbb84d6
+1.4.3 |checksum:b42a14f7e1d51a72e8f8a58748942ef86f1ba86f15e18d657035a87f6e056a67
+1.4.4 |checksum:bb7275a63396b55640653ebfef46aea626ceb716ddadff4c6be0221e5a143ccc
+1.4.5 |checksum:8b323fb4f60263d0ab19b7b1d92292fc0304afb4aba1c1b5da5182c871b8e3ac
+1.5.0 |checksum:dfe4a00007a710d238fb9f59b46336a4da1f54b4d90c23fde343959329f5ce20
+1.5.1 |checksum:170c448949e032cdbeed5ce466afa46a0eddc15c283ca6457fc64fe5a31829c2
+2.0.0.beta bson:~> 2.0.0.rc2,optionable:~> 0.1.1|checksum:78895153aa06ea6442c9ecc77fd8a15e7fd8d0412f77ab7989cc80a762bb6b98,rubygems:> 1.3.1
+2.0.0.beta2 bson:~> 2.0.0.rc2,optionable:~> 0.1.1|checksum:df435f8e1dca2cc37b9cdfb1b2e2348919c6b3781f761170fe6760ab1dec4b3c,rubygems:> 1.3.1
+2.0.0.beta3 bson:~> 2.0.0.rc3,optionable:~> 0.1.1|checksum:325fa0e94c12b98527067b21f5565f849642fa15ec61d429f0f492331fdc3cfd,rubygems:> 1.3.1
+2.0.0.beta4 bson:~> 2.0,connection_pool:~> 1.2,optionable:~> 0.2.0|checksum:2fa634e3bca63b2bd2ed279287182dc96a60c0da9f9d46144d9746273bb7f4dc,rubygems:> 1.3.1
+2.0.0.beta5 bson:~> 2.1,connection_pool:~> 1.2,optionable:~> 0.2.0|checksum:4002c37e73a99d7625769b01208d4c19fa4e321b765509310f7b990b563defca,rubygems:> 1.3.1
+1.5.2 |checksum:79348a87d65590d5470881e1173e12aac913f25c3654891c76329c191cae0970
+2.0.0.beta6 bson:~> 2.2,connection_pool:~> 1.2,optionable:~> 0.2.0|checksum:9d461f6ad9a8a9a5a059d31d59a17182ecab24d5ac2eb0553a94035d638e9540,rubygems:> 1.3.1
+2.0.0.rc1 bson:~> 2.2,connection_pool:~> 2.0,optionable:~> 0.2.0|checksum:a827b6ceaf8df8139074d19bdd84e8c11d56bfb3fdf80f2fcc0e9f618b655f31,rubygems:> 1.3.1
+2.0.0.rc2 bson:~> 2.2,connection_pool:~> 2.0,optionable:~> 0.2.0|checksum:002b9c7aa8f57c655d36f1c5e6797382238e367c292a8078136a4f982bee534d,rubygems:> 1.3.1
+2.0.0 bson:~> 2.2,connection_pool:~> 2.0,optionable:~> 0.2.0|checksum:7c4ca7603363d70b1af1437b080cf2269a7566fa26f140767d23a195eb216d77
+2.0.1 bson:~> 2.2,connection_pool:~> 2.0,optionable:~> 0.2.0|checksum:d7c07f9b38d1e2d4b9b701ff2803473f7c857f1605233f5c5cce0ec8a0747a72
+2.0.2 bson:~> 2.2,connection_pool:~> 2.0,optionable:~> 0.2.0|checksum:63676c5823b87f033f513310dd2f8bebabedb6b713677316c49de9a315d29c2a
+2.0.3 bson:~> 2.2,connection_pool:~> 2.0,optionable:~> 0.2.0|checksum:17244331cca4ebe7a7cdfcc7111ee9647f99997352af457b3ec5547b0c07aa7f
+2.0.4 bson:~> 2.2,connection_pool:~> 2.0,optionable:~> 0.2.0|checksum:d5cafc2b3110dd2b9a1bea115fbd546f6a0bad14751af85ffaeb9e37cf4676a9
+2.0.5 bson:~> 3.0.4,connection_pool:~> 2.0,optionable:~> 0.2.0|checksum:31c46b2bc5b6f58a341ae175e962f40b04d8f671f3378cfbd71c85844eb23061
+1.5.3 |checksum:44a3c3f8b698f80e41e64b978c60ec44d5def20bd24c7e26826b86f4d2994cb9
+2.0.6 bson:~> 3.0,connection_pool:~> 2.0,optionable:~> 0.2.0|checksum:c8c44bfef64598c3ca42691f2a52e8759ecd176b24ba7555d24be7a61fc77786
+2.0.7 bson:~> 3.0,connection_pool:~> 2.0,optionable:~> 0.2.0|checksum:e4931aa84f10796981edfcd0f5e53282018715c2926dd1849eb149646cf926fe