… bundler/lib/bundler/vendor/molinillo/lib/molinillo/resolution.rb
… bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/bson/GET/response
… bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/mongo/GET/response
… bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/mongoid/GET/response
… bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/optionable/GET/request
… bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/optionable/GET/response
… bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/origin/GET/request
… bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/origin/GET/response
… bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/will_paginate/GET/request
… bundler/spec/support/artifice/vcr_cassettes/realworld/index.rubygems.org/info/will_paginate/GET/response
… bundler/spec/support/artifice/vcr_cassettes/realworld/rubygems.org/gems/actionmailer-5.2.4.4.gem/GET/request
… bundler/spec/support/artifice/vcr_cassettes/realworld/rubygems.org/gems/actionmailer-5.2.4.4.gem/GET/response
… bundler/spec/support/artifice/vcr_cassettes/realworld/rubygems.org/gems/actionmailer-6.1.0.gem/GET/request
… bundler/spec/support/artifice/vcr_cassettes/realworld/rubygems.org/gems/actionmailer-6.1.0.gem/GET/response
… bundler/spec/support/artifice/vcr_cassettes/realworld/rubygems.org/gems/actionpack-5.2.4.4.gem/GET/request
… bundler/spec/support/artifice/vcr_cassettes/realworld/rubygems.org/gems/actionpack-5.2.4.4.gem/GET/response
… bundler/spec/support/artifice/vcr_cassettes/realworld/rubygems.org/gems/actionpack-6.1.0.gem/GET/request
… bundler/spec/support/artifice/vcr_cassettes/realworld/rubygems.org/gems/actionpack-6.1.0.gem/GET/response
… bundler/spec/support/artifice/vcr_cassettes/realworld/rubygems.org/gems/actionview-5.2.4.4.gem/GET/request
… bundler/spec/support/artifice/vcr_cassettes/realworld/rubygems.org/gems/actionview-5.2.4.4.gem/GET/response
… bundler/spec/support/artifice/vcr_cassettes/realworld/rubygems.org/gems/actionview-6.1.0.gem/GET/request
… bundler/spec/support/artifice/vcr_cassettes/realworld/rubygems.org/gems/actionview-6.1.0.gem/GET/response
… bundler/spec/support/artifice/vcr_cassettes/realworld/rubygems.org/gems/activejob-5.2.4.4.gem/GET/request
… bundler/spec/support/artifice/vcr_cassettes/realworld/rubygems.org/gems/activejob-5.2.4.4.gem/GET/response
… bundler/spec/support/artifice/vcr_cassettes/realworld/rubygems.org/gems/activejob-6.1.0.gem/GET/request
… bundler/spec/support/artifice/vcr_cassettes/realworld/rubygems.org/gems/activejob-6.1.0.gem/GET/response
… bundler/spec/support/artifice/vcr_cassettes/realworld/rubygems.org/gems/activemodel-5.2.4.4.gem/GET/request
… bundler/spec/support/artifice/vcr_cassettes/realworld/rubygems.org/gems/activemodel-5.2.4.4.gem/GET/response
… bundler/spec/support/artifice/vcr_cassettes/realworld/rubygems.org/gems/activesupport-5.2.4.4.gem/GET/request
… bundler/spec/support/artifice/vcr_cassettes/realworld/rubygems.org/gems/activesupport-5.2.4.4.gem/GET/response