<!--
Thanks so much for the contribution!

Note that you must abide by the [code of conduct](https://github.com/rubygems/rubygems/blob/master/CODE_OF_CONDUCT.md) to contribute to this project.

To make reviewing this PR a bit easier, please fill out answers to the following questions.
-->

## What was the end-user or developer problem that led to this PR?

<!-- Write a clear and complete description of the problem -->

Running some (relatively complex) resolution cases can lead to extremely poor performance of `Resolving dependencies`. Here is a trivial case to reproduce this which runs in 22 minutes on my laptop.

Gemfile
```ruby
source 'https://rubygems.org'

gem "actionmailer"
gem "mongoid", ">= 0.10.2"
```

The same problem has been seen in the real world at [Simply Business](https://www.simplybusiness.co.uk) where we have repeatedly run into resolution taking excessively long or even hanging.

## What is your fix for the problem, implemented in this PR?

<!-- Explain the fix being implemented. Include any diagnosis you run to
determine the cause of the issue and your conclusions. If you considered other
alternatives, explain why you end up choosing the current implementation -->

### Solution

The solution in this PR takes the above case from 22 minutes to 1 minute 32 seconds (on my laptop).

On other runs with different gem files, I've seen improvements including a 30% improvement and a 5 x improvement.

For the general case, there is no performance implications.

The basic solution (I will describe the diagnostics below) that led me to this requires two things:

1. Dedup `DepProxy` by creating a factory method and comparing requests for new `DepProxy`s to existing proxies (which creates a high cache hit rate). The using the hash of the objects to cache. This also allows comparison of two exactly the same `DepProxy`s to use the inbuilt `Object#==` which is very fast. This does mean that the objects are now shared. If they were immutable before, this would be totally fine, but even if they aren't, it appears not to cause problems.
2. Changing the Molinillo `build_details_for_unwind` in `resolution.rb` with an implementation for `relevant_unused_unwinds` that uses `Array#-` which takes advantage of object comparison rather than `Array#&` which uses `Array#hash` and in this case is substantially slower. Without the Molinillo change, the `DepProxy` factory is performance neutral.

`build_details_for_unwind` and specifically `Array#&` were where most of the time in the cases above were spent, as the analysis will demonstrate.

#### Note on the PR

I present the full PR with vendored changes, assuming if you would like to move forward I would submit a PR to `Molinillo`.

### Analysis

The added `slow_perf_spec.rb` can run against this branch or `master`. On this branch it runs in:

```
Finished in 4 minutes 37 seconds (files took 0.25304 seconds to load)
```

On `master` it runs:

```
Finished in 6 minutes 29 seconds (files took 0.27068 seconds to load)
```

This gives a savings of two minutes or ~ 30%.

Here is the output of the old and new algorithms:

#### DepProxy branch

```sh
$ rm Gemfile.lock; time ~/projects/rubygems/bundler/bin/bundle --gemfile=Gemfile
Fetching gem metadata from https://rubygems.org/........
Resolving dependencies...............
Using concurrent-ruby 1.1.7
Using minitest 5.14.2
Using zeitwerk 2.4.2
Using builder 3.2.4
Using erubi 1.10.0
Using mini_portile2 2.4.0
Using crass 1.0.6
Using rack 2.2.3
Using mini_mime 1.0.2
Using bson 4.11.1
Using bundler 2.3.0.dev
Using leshill-will_paginate 2.3.11
Using durran-validatable 2.0.1
Using i18n 1.8.5
Using nokogiri 1.10.10
Using tzinfo 2.0.4
Using loofah 2.8.0
Using activesupport 6.1.0
Using mongo 2.14.0
Using rails-html-sanitizer 1.3.0
Using rack-test 1.1.0
Using mail 2.7.1
Using globalid 0.4.2
Using rails-dom-testing 2.0.3
Using activejob 6.1.0
Using actionview 6.1.0
Using actionpack 6.1.0
Using mongoid 1.0.6
Using actionmailer 6.1.0
Bundle complete! 2 Gemfile dependencies, 29 gems now installed.
Use `bundle info [gemname]` to see where a bundled gem is installed.

real	1m31.824s
user	1m30.883s
sys	0m0.497s
```

#### Master branch

```sh
$ rm Gemfile.lock; time ~/projects/rubygems/bundler/bin/bundle --gemfile=Gemfile
Fetching gem metadata from https://rubygems.org/........
Resolving dependencies........................
Using concurrent-ruby 1.1.7
Using minitest 5.14.2
Using zeitwerk 2.4.2
Using builder 3.2.4
Using erubi 1.10.0
Using mini_portile2 2.4.0
Using crass 1.0.6
Using rack 2.2.3
Using mini_mime 1.0.2
Using bson 4.11.1
Using bundler 2.3.0.dev
Using leshill-will_paginate 2.3.11
Using durran-validatable 2.0.1
Using i18n 1.8.5
Using tzinfo 2.0.4
Using nokogiri 1.10.10
Using activesupport 6.1.0
Using mail 2.7.1
Using rails-dom-testing 2.0.3
Using rack-test 1.1.0
Using globalid 0.4.2
Using mongo 2.14.0
Using loofah 2.8.0
Using activejob 6.1.0
Using rails-html-sanitizer 1.3.0
Using mongoid 1.0.6
Using actionview 6.1.0
Using actionpack 6.1.0
Using actionmailer 6.1.0
Bundle complete! 2 Gemfile dependencies, 29 gems now installed.
Use `bundle info [gemname]` to see where a bundled gem is installed.

real	22m27.811s
user	22m15.555s
sys	0m4.723s
```

#### The results of running `ruby-prof` on `master` and on the new branch

##### Original implementation

```
Measure Mode: wall_time
Thread ID: 70207129452500
Fiber ID: 70207121129620
Total: 514.729486
Sort by: self_time

 %self      total      self      wait     child     calls  name                           location
 54.20    481.506   278.970     0.000   202.536  2230632   Array#&
 38.29    197.141   197.101     0.000     0.039 870039195   Bundler::DepProxy#hash         /Users/lukas.oberhuber/projects/rubygems/bundler/lib/bundler/dep_proxy.rb:12
  0.95     12.820     4.877     0.000     7.943  7039049   Bundler::DepProxy#eql?         /Users/lukas.oberhuber/projects/rubygems/bundler/lib/bundler/dep_proxy.rb:16
  0.72    493.137     3.682     0.000   489.455   233778  *Array#select
  0.65      7.320     3.367     0.000     3.953  8345705   Gem::Dependency#==             /Users/lukas.oberhuber/.rbenv/versions/2.6.5/lib/ruby/2.6.0/rubygems/dependency.rb:173
  0.37      2.953     1.899     0.000     1.054  2239519   Bundler::Molinillo::Resolver::Resolution::UnwindDetails#<=> /Users/lukas.oberhuber/projects/rubygems/bundler/lib/bundler/vendor/molinillo/lib/molinillo/resolution.rb:83
  0.37     10.352     1.884     0.000     8.469   138522  *Array#each
  0.36      1.876     1.876     0.000     0.000 15642359   Bundler::Molinillo::Resolver::Resolution::UnwindDetails#state_index
  ...
```

##### With adding a DepProxy factory

```
Measure Mode: wall_time
Thread ID: 70284950585360
Fiber ID: 70284980000360
Total: 97.786107
Sort by: self_time

 %self      total      self      wait     child     calls  name                           location
 76.78     75.077    75.077     0.000     0.000  2327396   Array#-
  3.54     85.954     3.465     0.000    82.489    95397   Array#select
  1.87      2.805     1.831     0.000     0.974  2239519   Bundler::Molinillo::Resolver::Resolution::UnwindDetails#<=> /Users/lukas.oberhuber/projects/rubygems/bundler/lib/bundler/vendor/molinillo/lib/molinillo/resolution.rb:83
  1.83      1.789     1.789     0.000     0.000 15642359   Bundler::Molinillo::Resolver::Resolution::UnwindDetails#state_index
  1.81      9.044     1.774     0.000     7.270   138532  *Array#each
  0.95      1.570     0.928     0.000     0.642   260374   Gem::Version#<=>               /Users/lukas.oberhuber/.rbenv/versions/2.6.5/lib/ruby/2.6.0/rubygems/version.rb:342
  0.87      0.852     0.852     0.000     0.000  4454118   Bundler::Molinillo::Resolver::Resolution::UnwindDetails#requirements_unwound_to_instead
  0.74      3.494     0.720     0.000     2.774  2218106   Comparable#>
  0.66      1.028     0.650     0.000     0.378   350426   Array#reject!
  ...
```

### Other interesting findings

Interestingly, by switching the `mongoid` version from `>=0.10.2` to `>=0.10.3`, steps to resolve go from 6264 down to 827 due to picking a different starting requirment (although I haven't analysed the algorithm to understand why the change is triggered). Red or `<` version is `>=0.10.3` and Green or `>` is `>=0.10.2`. There is also a dramatic slowdown in dependency resolution between the two options, due to the very slow algorithm (even with the fixes proposed here) related to looking for previous unwind states.

```diff
< BUNDLER: Starting resolution (2020-12-30 00:44:11 +0000)
< BUNDLER: User-requested dependencies: [#<Bundler::DepProxy:0x00007fef5f0d8058 @dep=<Bundler::Dependency type=:runtime name="actionmailer" requirements=">= 0">, @__platform=#<Gem::Platform:0x00007fef5f016098 @cpu="x86_64", @os="darwin", @version="19" @cpu="x86_64", @os="darwin", @version="19">, @hash=325717142915182719>, #<Bundler::DepProxy:0x00007fef66073408 @dep=<Bundler::Dependency type=:runtime name="mongoid" requirements=">= 0.10.3">, @__platform=#<Gem::Platform:0x00007fef5f016098 @cpu="x86_64", @os="darwin", @version="19" @cpu="x86_64", @os="darwin", @version="19">, @hash=1448836936842474097>, #<Bundler::DepProxy:0x00007fef66073070 @dep=<Bundler::Dependency type=:runtime name="Ruby\u0000" requirements=">= 0">, @__platform=#<Gem::Platform:0x00007fef5f016098 @cpu="x86_64", @os="darwin", @version="19" @cpu="x86_64", @os="darwin", @version="19">, @hash=-4097237829671024793>, #<Bundler::DepProxy:0x00007fef66072800 @dep=<Bundler::Dependency type=:runtime name="RubyGems\u0000" requirements="= 3.1.4">, @__platform=#<Gem::Platform:0x00007fef5f016098 @cpu="x86_64", @os="darwin", @version="19" @cpu="x86_64", @os="darwin", @version="19">, @hash=-3847635433202042685>]
---
> BUNDLER: Starting resolution (2020-12-30 00:40:45 +0000)
> BUNDLER: User-requested dependencies: [#<Bundler::DepProxy:0x00007fc6070f4518 @dep=<Bundler::Dependency type=:runtime name="actionmailer" requirements=">= 0">, @__platform=#<Gem::Platform:0x00007fc6078f1ec8 @cpu="x86_64", @os="darwin", @version="19" @cpu="x86_64", @os="darwin", @version="19">, @hash=-3111283843832305287>, #<Bundler::DepProxy:0x00007fc60c147dd0 @dep=<Bundler::Dependency type=:runtime name="mongoid" requirements=">= 0.10.2">, @__platform=#<Gem::Platform:0x00007fc6078f1ec8 @cpu="x86_64", @os="darwin", @version="19" @cpu="x86_64", @os="darwin", @version="19">, @hash=-4225324059996933406>, #<Bundler::DepProxy:0x00007fc60c147358 @dep=<Bundler::Dependency type=:runtime name="Ruby\u0000" requirements=">= 0">, @__platform=#<Gem::Platform:0x00007fc6078f1ec8 @cpu="x86_64", @os="darwin", @version="19" @cpu="x86_64", @os="darwin", @version="19">, @hash=-329411114533298518>, #<Bundler::DepProxy:0x00007fc60c146ed0 @dep=<Bundler::Dependency type=:runtime name="RubyGems\u0000" requirements="= 3.1.4">, @__platform=#<Gem::Platform:0x00007fc6078f1ec8 @cpu="x86_64", @os="darwin", @version="19" @cpu="x86_64", @os="darwin", @version="19">, @hash=-3561920153128883523>]
13,20c13,20
< BUNDLER:     Creating possibility state for mongoid (>= 0.10.3) x86_64-darwin-19 (428 remaining)
< BUNDLER:       Attempting to activate [mongoid (7.2.0) (x86_64-darwin-19)]
< BUNDLER:       Activated mongoid at [mongoid (7.2.0) (x86_64-darwin-19)]
< BUNDLER:       Requiring nested dependencies (activemodel (>= 5.1, < 6.1) x86_64-darwin-19, mongo (>= 2.10.5, < 3.0.0) x86_64-darwin-19)
< BUNDLER:       Creating possibility state for actionmailer x86_64-darwin-19 (744 remaining)
< BUNDLER:         Attempting to activate [actionmailer (6.1.0) (x86_64-darwin-19)]
< BUNDLER:         Activated actionmailer at [actionmailer (6.1.0) (x86_64-darwin-19)]
< BUNDLER:         Requiring nested dependencies (activesupport (= 6.1.0) x86_64-darwin-19, actionpack (= 6.1.0) x86_64-darwin-19, actionview (= 6.1.0) x86_64-darwin-19, activejob (= 6.1.0) x86_64-darwin-19, mail (~> 2.5, >= 2.5.4) x86_64-darwin-19, rails-dom-testing (~> 2.0) x86_64-darwin-19)
---
> BUNDLER:     Creating possibility state for actionmailer x86_64-darwin-19 (744 remaining)
> BUNDLER:       Attempting to activate [actionmailer (6.1.0) (x86_64-darwin-19)]
> BUNDLER:       Activated actionmailer at [actionmailer (6.1.0) (x86_64-darwin-19)]
> BUNDLER:       Requiring nested dependencies (activesupport (= 6.1.0) x86_64-darwin-19, actionpack (= 6.1.0) x86_64-darwin-19, actionview (= 6.1.0) x86_64-darwin-19, activejob (= 6.1.0) x86_64-darwin-19, mail (~> 2.5, >= 2.5.4) x86_64-darwin-19, rails-dom-testing (~> 2.0) x86_64-darwin-19)
> BUNDLER:       Creating possibility state for mongoid (>= 0.10.2) x86_64-darwin-19 (430 remaining)
> BUNDLER:         Attempting to activate [mongoid (7.2.0) (x86_64-darwin-19)]
> BUNDLER:         Activated mongoid at [mongoid (7.2.0) (x86_64-darwin-19)]
> BUNDLER:         Requiring nested dependencies (activemodel (>= 5.1, < 6.1) x86_64-darwin-19, mongo (>= 2.10.5, < 3.0.0) x86_64-darwin-19)
```

## Make sure the following tasks are checked

- [x ] Describe the problem / feature
- [x ] Write [tests](https://github.com/rubygems/rubygems/blob/master/bundler/doc/development/PULL_REQUESTS.md#tests) for features and bug fixes
- [x ] Write code to solve the problem
- [x ] Make sure you follow the [current code style](https://github.com/rubygems/rubygems/blob/master/bundler/doc/development/PULL_REQUESTS.md#code-formatting) and [write meaningful commit messages without tags](https://github.com/rubygems/rubygems/blob/master/bundler/doc/development/PULL_REQUESTS.md#commit-messages)