@lukaso 2020-12-31
This additional change looks like it took this case from 1m32s down to 10s on my laptop. The `requirements_unwound_to_instead` array was not being deduplicated on insert.