@deivid-rodriguez 2021-01-07
Ok, I've been working on finishing this up and it's almost ready. I'll force-push to this branch soon.

Current performance on my machine with the problematic Gemfile is ~3 seconds without the molinillo patch, and ~10 seconds without it :tada:.