@deivid-rodriguez 2021-01-07
Great! I created a PR to Molinillo: https://github.com/CocoaPods/Molinillo/pull/146. If it passes Molinillo's specs I think we can ship it.