@deivid-rodriguez 2021-01-07
Yay, it's great! I added some tweaks to get this ready. Suming them up:

* Removed the changes to the realworld specs, since I wasn't sure. I'll still get specs working on MacOS on a separate PR.
* Given that the spec now takes 3 seconds, I cut the limit to fail the realworld spec to 120 seconds.
* I tweaked the fix in #4250 here, added a "polyfill" to `bundler` to support older rubygems not including the fix, and removed the tweak in hashes from `DepProxy`.
* I added a note in the Rakefile to mention that we're cherry-picking some unreleased changes to Molinillo.
* I regenerated cassettes, given recent changes from other PRs.

I think that's it.

@simi You can have a look now!