@segiddins 2021-01-07
Because it’s mapping from an infinite dimensional space (any possible array) to a fixed-width integer. 

On Jan 7, 2021, at 12:58 PM, David Rodríguez <notifications@github.com> wrote:

﻿
@deivid-rodriguez commented on this pull request.

In bundler/lib/bundler/dep_proxy.rb:

> @@ -4,19 +4,18 @@ module Bundler
   class DepProxy
     attr_reader :__platform, :dep
 
+    @proxies = {}
+
+    def self.get_proxy(dep, platform)
+      @proxies[[dep, platform].hash] ||= new(dep, platform).freeze
I think we are confusing the #hash (method) ruby uses to determine the key used under the hood when using objects as keys with the hash (function) used by the hash (data structure) implementation to map keys to buckets. The hash (function) can certainly have collisions that should be handled by the implementation, but I don't see why the #hash method on an Array should have collisions and not be unique to the array contents.

—
You are receiving this because you were mentioned.
Reply to this email directly, view it on GitHub, or unsubscribe.