@deivid-rodriguez 2021-01-07
That makes sense @segiddins, but won't ruby take care of collisions when generating hashes? If you implement your own `#hash` you need to make sure that you can't have collisions, but I would expect ruby's `Array#hash` to not generate collisions? At least the docs seem to suggest so.

In any case, even if we were to really fix this I'm not sure how to go about it, since using an array as the hash key I would imagine `ruby` would use `Array#hash` under the hood for the key, and with my suggestion we still end up using `Gem::Platform` and `Gem::Requirement` as keys, which use `Array#hash` to implement their own `#hash` methods.

So, to sum up, I don't think this is a problem in practice so I'm fine with whatever alternative you or @indirect think is safest. I'll just update this to use the `Array` as the key as you suggested.