@deivid-rodriguez 2021-01-11
@simi In case it helps with your review (assuming you're interested in reviewing the part about hash collisions specifically), I made some research on ruby sources:

This defines `Hash#[]`: https://github.com/ruby/ruby/blob/55e52c19e74d5df90560ea1cc4f2a2b9f5d7a5c4/hash.c#L6874, which is this method: https://github.com/ruby/ruby/blob/55e52c19e74d5df90560ea1cc4f2a2b9f5d7a5c4/hash.c#L2065, which calls the function `hash_st_like_lookup` and ends up calling `st_lookup` which implements the juicy part: https://github.com/ruby/ruby/blob/55e52c19e74d5df90560ea1cc4f2a2b9f5d7a5c4/st.c#L1007-L1031. I think the `do_hash` call in the second line calls `#hash` on the given key, in particular  this line: https://github.com/ruby/ruby/blob/55e52c19e74d5df90560ea1cc4f2a2b9f5d7a5c4/st.c#L320.

If my interpretation of the C code is correct (which is assuming too much already :sweat_smile:), that would mean that all alternatives proposed here are equally safe (or unsafe, although I would say they are all safe).