@lukaso 2021-01-11
> If my interpretation of the C code is correct (which is assuming too much already 😅), that would mean that all alternatives proposed here are equally safe (or unsafe, although I would say they are all safe).

I think the `[dep, platform]` without the `.hash` is best. It uses ruby's hash key functionality which does not allow hash collisions. I ran a test and there are many cache hits with this variant, meaning two instances of `dep` that are equal evaluate to the same key. Given this, the current implementation on the branch is the best from my reading and testing.