@deivid-rodriguez 2021-01-13
Still need to investigate it more, but as of today, the `Gemfile` introduced as part of #4134 is again unbearably slow (I aborted it after 18 minutes).

It seems that some recent gem releases made the number of steps in that resolution go from ~1000 to ~6000. With this PR, resolution is at least workable taking ~ 1minute.

All that to say, we should merge this. I think it has gone through enough scrutiny already. If nobody beats me to it, I'll be merging this later today, we can address any late maintainer feedback post-merge.